take MyGUIEngine fixes and features from trunk (Part 1)

+++ branches/trunk_3.0/MyGUIEngine/src/MyGUI_HScroll.cpp
@@ -79,7 +79,7 @@ namespace MyGUI
 		{
 			mWidgetTrack->setVisible(false);
 			if ( nullptr != mWidgetFirstPart ) mWidgetFirstPart->setSize(pos/2, mWidgetFirstPart->getHeight());
-			if ( nullptr != mWidgetSecondPart ) mWidgetSecondPart->setCoord(pos/2 + mSkinRangeStart, mWidgetSecondPart->getTop(), pos - pos/2, mWidgetSecondPart->getHeight());
+			if ( nullptr != mWidgetSecondPart ) mWidgetSecondPart->setCoord(pos/2 + (int)mSkinRangeStart, mWidgetSecondPart->getTop(), pos - pos/2, mWidgetSecondPart->getHeight());
 			return;
 		}
 		// если скрыт то покажем
